Yk/refine

+++ ConnectedVehicles-Customer/src/main/java/com/youssef/connectedvehicles/service/SequenceGeneratorService.java
@@ -21,7 +21,7 @@ public class SequenceGeneratorService {
 		this.mongoOperations = mongoOperations;
 	}
 
-	public long generateSequence(String seqName) {
+	public int generateSequence(String seqName) {
 		DatabaseSequence counter = mongoOperations.findAndModify(
 				query(where("_id").is(seqName)),
 				new Update().inc("seq", 1),
